Diagrama y avaze de microservicio

+++ paper_microservice/Paper/src/main/java/co/edu/unicauca/cuychair/paper_microservice/layerservices/DTO/PaperDTO.java
@@ -11,8 +11,8 @@ public class PaperDTO {
     private int id;
     private String title;
     private String description;
-    private Long idAuthor;
-    private Long idConference;
+    private int idAuthor;
+    private int idConference;
 
     public PaperDTO() {
     }
